ai/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.7.2
pydantic>=2.7.1
python-dotenv>=1.0.1
PyJWT>=2.8.0
spacy>=3.8.0
transformers>=4.40.0
torch>=2.4.0
httpx>=0.27.0
python-multipart>=0.0.9
